requirements.txt
# Core
grpcio>=1.76.0,<2.0.0
grpcio-tools>=1.76.0,<2.0.0
protobuf>=6.31.1,<7.0.0
numpy>=1.26.0,<=2.3
psutil>=5.9.0
requests>=2.31.0
httpx[http2]>=0.27.0
tenacity>=8.2.0
rich>=13.6.0
pydantic>=2.4.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
setuptools<70.0.0
numba>=0.57.0
cryptography>=42.0.0
pycryptodome>=3.20.0

# Media Processing
yt-dlp>=2023.10.13
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
webrtcvad>=2.0.10

# AI & ML
faster-whisper>=0.9.0
openai>=1.3.0
dashscope>=1.14.0
sentence-transformers>=2.2.2
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
paddlepaddle==3.3.0; sys_platform == 'win32'
paddleocr==2.7.3
paddlex==3.4.2
pytesseract>=0.3.10
mediapipe==0.10.14

# Testing
pytest>=7.4.0

# Graph & Tools
langgraph>=0.0.10
requests-toolbelt>=1.0.0
